+++ api/do_test.go
@@ -1,0 +1,28 @@
+package api_test
+
+import (
+	"testing"
+
+	"github.com/machbase/neo-server/api"
+	"github.com/stretchr/testify/require"
+)
+
+func TestTableName(t *testing.T) {
+	tests := []struct {
+		input  string
+		expect [3]string
+	}{
+		{"a.b.c", [3]string{"A", "B", "C"}},
+		{"user.table", [3]string{"MACHBASEDB", "USER", "TABLE"}},
+		{"table", [3]string{"MACHBASEDB", "SYS", "TABLE"}},
+	}
+
+	for _, test := range tests {
+		t.Run(test.input, func(t *testing.T) {
+			a, b, c := api.TableName(test.input).Split()
+			require.Equal(t, test.expect[0], a)
+			require.Equal(t, test.expect[1], b)
+			require.Equal(t, test.expect[2], c)
+		})
+	}
+}
